requirements.txt
asgiref==3.2.3
certifi==2019.11.28
Django==3.0.2
django-ckeditor==5.8.0
django-js-asset==1.2.2
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
